scripts/requirements.txt
# Data pipeline dependencies
httpx==0.28.1
python-dotenv==1.0.1
